Move handler functions to helper folder for code readability
