+++ src/components/OrderTable/OrderTable.js
@@ -16,6 +16,7 @@ const OrderTable = ({
       </thead>
       <tbody>
         {Object.keys(cart).map((productId) => {
+          console.log(products);
           const product = products.find(
             (p) => p.id === parseInt(productId)
           );
