+++ src/helpers/helpers.js
@@ -1,0 +1,38 @@
+export const addOrRemoveProductFromCart = (setCart) => {
+  const addProductToCart = (productId) => {
+    setCart((prevCart) => {
+      const newCart = { ...prevCart };
+      newCart[productId] = (newCart[productId] || 0) + 1;
+      return newCart;
+    });
+  };
+
+  const removeProductFromCart = (productId) => {
+    setCart((prevCart) => {
+      const newCart = { ...prevCart };
+      if (newCart[productId] > 0) {
+        newCart[productId] -= 1;
+      }
+      if (newCart[productId] === 0) {
+        delete newCart[productId];
+      }
+      return newCart;
+    });
+  };
+
+  return { addProductToCart, removeProductFromCart };
+};
+
+export const totalPrice = (cart, products) => {
+  const cartSubtotal = Object.keys(cart).reduce(
+    (acc, productId) =>
+      acc +
+      products.find((p) => p.id === parseInt(productId)).price *
+        cart[productId],
+    0
+  );
+  const discount = cartSubtotal > 1000 ? cartSubtotal * 0.1 : 0;
+  const total = cartSubtotal - discount;
+
+  return { cartSubtotal, discount, total };
+};
